gas gas gas

+++ src/extract.py
@@ -1,63 +1,48 @@
-from pdfminer.high_level import extract_text
+from pdfminer.high_level    import extract_text
+from dotenv                 import load_dotenv
+from pathlib                import Path
+from itertools              import chain
 import re
-from itertools import chain
-import openai
 import json
-from dotenv import load_dotenv
-from pathlib import Path
-import os
-env_path = Path('.')/'.env'
-load_dotenv(dotenv_path=env_path)
-openai.api_key = token=os.environ['CHAT_TOKEN']
 
-def _get_sentences(file_path: str) -> list[str]:
-    text = extract_text(file_path)
-    sentences = re.split(r'[.!?]', text)
-    sentences = [sentence.strip() for sentence in sentences if sentence.strip()]
-    return sentences
+import openai
+import os
+
+env_path =                  Path('.')/'.env'
+load_dotenv(dotenv_path=    env_path)
+openai.api_key =            token=os.environ['CHAT_TOKEN']
+
 def printlist(arr: list):
     for i, element in enumerate(arr):
         print(i)
         print(element) 
         print("--------------------------------\n\n\n\n")
-def _get_windows(sentences: list[str], max_words:int =4000, overlap:int=200) -> list[str]:
-    words = list(chain(*[sentence.split() for sentence in sentences]))
-    start_idx = 0
 
-    while start_idx < len(words):
-        end_idx = start_idx + max_words
-        if end_idx > len(words):
-            end_idx = len(words)
-
-        window_words = words[start_idx:end_idx]
-        window_text = ' '.join(window_words)
-        yield window_text
-        start_idx += (max_words - overlap)
 def get_windows(sentences: list[str], max_words:int =4000, overlap:int=200) -> list[str]:
-    words = list(chain(*[sentence.split() for sentence in sentences]))
-    start_idx = 0
-
-    while start_idx < len(words):
-        end_idx = start_idx + max_words
-        if end_idx > len(words):
-            end_idx = len(words)
-
-        window_words = words[start_idx:end_idx]
-        window_text = ' '.join(window_words)
+    words =                     list(chain(*[sentence.split() for sentence in sentences]))
+    start_idx =                 0
+    while   start_idx < len(words):
+        end_idx =               start_idx + max_words
+        if  end_idx > len(words):
+            end_idx =           len(words)
+        window_words =          words[start_idx:end_idx]
+        window_text =           ' '.join(window_words)
         yield window_text
-        start_idx += (max_words - overlap)
+        start_idx +=            (max_words - overlap)
+
 def get_sentences(file_path: str) -> list[str]:
-    text = extract_text(file_path)
-    sentences = re.split(r'[.!?]', text)
-    sentences = [sentence.strip() for sentence in sentences if sentence.strip()]
+    text =                      extract_text(file_path)
+    sentences =                 re.split(r'[.!?]', text)
+    sentences =                 [sentence.strip() for sentence in sentences if sentence.strip()]
     return sentences
+
 def extract_points(chunk, topics):
     system_prompt = (
         "You are an interviewer trying to extract the main points out from a piece of paper in preperation for an interview. "
         "Use the extract() function to get the chapter names and keypoints out from the paper. The keys for the json are: topics: [topic, subtopics : [subtopic]]"   
     )
-    user_prompt = "Text to extract content from:\n"
-    user_prompt += chunk
+    user_prompt =               "Text to extract content from:\n"
+    user_prompt +=              chunk
     functions = [
         {
             "name": "extract_topics",
@@ -97,68 +82,72 @@ def extract_points(chunk, topics):
         },
     ]
     completion = openai.ChatCompletion.create(
-        model = "gpt-3.5-turbo-16k",
+        model =             "gpt-3.5-turbo-16k",
         messages = [
-            {"role": "system", "content": system_prompt},
-            {"role": "user", "content": user_prompt},
-
+                            {"role": "system", "content": system_prompt},
+                            {"role": "user", "content": user_prompt},
         ],
-        functions = functions,
-        function_call = {"name": "extract_topics"},
-        max_tokens=5000
+        functions =         functions,
+        function_call =     {"name": "extract_topics"},
+        max_tokens =        5000
     )
-    choice = completion.choices[0]
-    encoded_data = choice.message.function_call.arguments
+    choice =                completion.choices[0]
+    print(choice)
+    encoded_data =          choice.message.function_call.arguments
     #print(json.dumps(json.loads(encoded_data), indent=4))
     return encoded_data
 
 def extract_chunk(chunk, extraction_count:int = 5):
-    extractions = []
+    extractions =           []
     for i in range(extraction_count): 
-        topics = []
+        topics =            []
         #SWAP OUT EXTRACTION FUNCTION HERE
         #---------------------------------
-        json_raw = extract_points(chunk, topics)
-        print(json_raw)
-        data = json.loads(json_raw)
-        topics_key = list(data.keys())
-        topics = data[topics_key[0]]
-        topic_count = len(topics) 
-        subtopic_count = 0
+        json_raw =          extract_points(chunk, topics)
+        #print(json_raw)
+        data =              json.loads(json_raw)
+        topics_key =        list(data.keys())
+        topics =            data[topics_key[0]]
+        topic_count =       len(topics) 
+        subtopic_count =    0
         for topic in topics:
             try:
-                topic_keys = list(topic.keys())
+                topic_keys =            list(topic.keys())
                 try:
-                    subtopic_key = topic_keys[1]
-                    subtopic_count += len(topic[subtopic_key])
+                    subtopic_key =      topic_keys[1]
+                    subtopic_count +=   len(topic[subtopic_key])
                 except:
                     pass
             except:
                 pass
         #print("Topics: " + str(topic_count))
         #print("Subtopics: " + str(subtopic_count))
-        extraction_score = topic_count + subtopic_count*0.5
-        extraction = [topics, topic_count, subtopic_count, extraction_score]
+        extraction_score =              topic_count + subtopic_count*0.5
+        extraction =                    [topics, topic_count, subtopic_count, extraction_score]
         extractions.append(extraction)
-    sorted_extractions = sorted(extractions, key=lambda x: x[3], reverse=True)
-    best_extraction = sorted_extractions[0]
+    sorted_extractions =                sorted(extractions, key=lambda x: x[3], reverse=True)
+    best_extraction =                   sorted_extractions[0]
     return best_extraction
+
 def extract_doc(filepath, savepath):
-    sentences = _get_sentences(filepath)
-    windows = list(_get_windows(sentences))
+    if os.path.exists(savepath):
+        return True
+    sentences =             get_sentences(filepath)
+    windows =               list(get_windows(sentences))
     print(len(windows))
-    extractions = []
-    for i, chunk  in enumerate(windows):
+    extractions =           []
+    for i, chunk  in enumerate(windows[:3]):
         print(i)
-        extraction = extract_chunk(str(chunk))
+        extraction =        extract_chunk(str(chunk))
         extraction.append(i)
         extractions.append(extraction)
-    data = {"data": extractions}
+    data =                  {"data": extractions}
     with open(savepath, "w") as f:
         json.dump(data, f)
+
 def load_extraction(filepath):
     with open(filepath, "r") as f:
-        data = json.load(f)
+        data =              json.load(f)
     return data
 
 
@@ -174,6 +163,6 @@ if __name__ == "__main__":
         "VOYAGER: An Open-Ended Embodied Agent with Large Language Models",
     ]
     paper_index = 2
-    paper = [paper_index]
-    filepath = file_paths[paper_index]
+    paper =                 [paper_index]
+    filepath =              file_paths[paper_index]
     extract_doc(filepath, "./cache/voyager.json")
